feat: refactor

+++ db/schema/001_youtube.sql
@@ -12,10 +12,10 @@ CREATE TABLE IF NOT EXISTS videos (
     thumbnail_url TEXT NOT NULL,
     provider TEXT NOT NULL, -- provides details about video provider 'youtube', 'vimeo', 'facebook' etc
     video_id TEXT UNIQUE NOT NULL, -- contains information about the actual url of the video from where we can stream it
-    view_count INT NOT NULL,
-    like_count INT NOT NULL,
-    favorite_count INT NOT NULL,
-    comment_count INT NOT NULL
+    view_count BIGINT NOT NULL,
+    like_count BIGINT NOT NULL,
+    favorite_count BIGINT NOT NULL,
+    comment_count BIGINT NOT NULL
 );
